Remove platform specific index files
The method declarations in index.js were empty,
therefore I had to import it from either .ios or .android
which is not that convenient.
Fixes issue #74.
